sensors and stuff

+++ backend/scripts/generateManyToManyBoilerplate/index.js
@@ -4,7 +4,7 @@ import { camelCase, pascalCase, snakeCase } from 'change-case';
 import prompts from 'prompts';
 import { appendFile, writeFile } from 'fs/promises';
 
-const { first, second } = await prompts([
+const { first, second, dryRun } = await prompts([
   {
     type: 'text',
     name: 'first',
@@ -14,6 +14,14 @@ const { first, second } = await prompts([
     type: 'text',
     name: 'second',
     message: 'Second entity name (fully lower case) with space delimiter',
+  },
+  {
+    type: 'toggle',
+    name: 'dryRun',
+    message: 'Dry run? (skip real writes to file?)',
+    initial: false,
+    active: 'yes',
+    inactive: 'no',
   },
 ]);
 
@@ -33,7 +41,7 @@ import { ${secondPascal}, ${firstPascal} } from '.';
 export class ${firstPascal}To${secondPascal} {
   @ManyToOne(
     () => ${firstPascal},
-    (${firstCamel}) => ${firstCamel}.${firstCamel}To${secondPascal}Relations
+    (${firstCamel}) => ${firstCamel}.${firstCamel}To${secondPascal}Relations,
   )
   @JoinColumn({ name: '${firstSnake}_id' })
   ${firstCamel}!: ${firstPascal};
@@ -57,6 +65,20 @@ export class ${firstPascal}To${secondPascal} {
     primary: true,
     nullable: false,
   })
+  ${secondCamel}Id!: number;
+}
+`;
+
+const getIntermediateModelInterface =
+  () => `import { I${secondPascal}, I${firstPascal} } from '.';
+
+export class I${firstPascal}To${secondPascal} {
+  ${firstCamel}!: I${firstPascal};
+
+  ${firstCamel}Id!: number;
+
+  ${secondCamel}!: I${secondPascal};
+
   ${secondCamel}Id!: number;
 }
 `;
@@ -86,7 +108,7 @@ const getFirstModelMixin = () => `
 const getSecondModelMixin = () => `
   @ManyToMany(
     () => ${firstPascal},
-    (${firstCamel}) => ${firstCamel}.${secondCamel}s
+    (${firstCamel}) => ${firstCamel}.${secondCamel}s,
   )
   ${firstCamel}sWithThat${secondPascal}!: ${firstPascal}[];
 
@@ -97,18 +119,54 @@ const getSecondModelMixin = () => `
   ${firstCamel}To${secondPascal}Relations!: ${firstPascal}To${secondPascal}[];
 `;
 
+const getFirstModelInterfaceMixin = () => `
+  ${secondCamel}s!: I${secondPascal}[];
+
+  ${firstCamel}To${secondPascal}Relations!: I${firstPascal}To${secondPascal}[];
+`;
+
+const getSecondModelInterfaceMixin = () => `
+  ${firstCamel}sWithThat${secondPascal}!: I${firstPascal}[];
+
+  ${firstCamel}To${secondPascal}Relations!: I${firstPascal}To${secondPascal}[];
+`;
+
 console.log(`new ${firstPascal}To${secondPascal} model was generated\n`);
 console.log(getIntermediateModel());
-await writeFile(
-  `./backend/src/modules/infrastructure/model/${firstCamel}To${secondPascal}.model.ts`,
-  getIntermediateModel(),
-);
-await appendFile(
-  `./backend/src/modules/infrastructure/model/index.ts`,
-  `export * from './${firstCamel}To${secondPascal}.model';\n`,
-);
+
+if (!dryRun) {
+  await writeFile(
+    `./backend/src/modules/infrastructure/model/${firstCamel}To${secondPascal}.model.ts`,
+    getIntermediateModel(),
+  );
+  await appendFile(
+    `./backend/src/modules/infrastructure/model/index.ts`,
+    `export * from './${firstCamel}To${secondPascal}.model';\n`,
+  );
+}
 
 console.log(`\n------ Mixin for ${firstPascal} model:\n`);
 console.log(getFirstModelMixin());
 console.log(`\n------ Mixin for ${secondPascal} model:\n`);
 console.log(getSecondModelMixin());
+
+console.log(
+  `\n------ new I${firstPascal}To${secondPascal} model interface was generated:\n`,
+);
+console.log(getIntermediateModelInterface());
+
+if (!dryRun) {
+  await writeFile(
+    `./shared/src/model/${firstCamel}To${secondPascal}.model.ts`,
+    getIntermediateModelInterface(),
+  );
+  await appendFile(
+    `./shared/src/model/index.ts`,
+    `export * from './${firstCamel}To${secondPascal}.model';\n`,
+  );
+}
+
+console.log(`\n------ Mixin for I${firstPascal} model interface:\n`);
+console.log(getFirstModelInterfaceMixin());
+console.log(`\n------ Mixin for I${secondPascal} model interface:\n`);
+console.log(getSecondModelInterfaceMixin());
